+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog][keepachangelog] and this project adheres to [Semantic Versioning][semver].
 
+
+### Added
+
+- Reference with flags of the initiator of the transaction
+
 ## v1.7.0
 
 ### Added